Fixed units for MatchScore, MatcheeHeight & MatcheeWeight

+++ app/src/main/java/com/example/cs446_fit4me/model/MatchingModel.kt
@@ -5,7 +5,7 @@ enum class GymFrequency { NEVER, RARELY, OCCASIONALLY, REGULARLY, FREQUENTLY, DA
 data class MatchEntry(
     val userId: String,
     val matchedUserId: String,
-    val score: Int,
+    val score: Float,
     val matchee: MatcheeUser? // You'll need to define this model as well
 )
 
@@ -13,8 +13,8 @@ data class MatcheeUser(
     val id: String,
     val name: String,
     val age: Int,
-    val heightCm: Int,
-    val weightKg: Int,
+    val heightCm: Float,
+    val weightKg: Float,
     val location: String,
     val timePreference: TimePreference,
     val experienceLevel: ExperienceLevel,
